docker/requirements/backend/requirements.txt
asgiref==3.7.2
Django==3.2.25
pytz==2024.2
sqlparse==0.4.4
typing_extensions==4.7.1
gunicorn==23.0.0
psycopg2-binary
django-cors-headers==4.4.0
requests==2.32.3
djangorestframework==3.13.1